fix onTokenTransfer

+++ evm-contracts/src/v0.4/interfaces/OracleInterface.sol
@@ -13,5 +13,5 @@ interface OracleInterface {
   function setFulfillmentPermission(address node, bool allowed) external;
   function withdraw(address recipient, uint256 amount) external;
   function withdrawable() external view returns (uint256);
-  function onTokenTransfer(address _sender, uint _value, bytes _data);
+  function onTokenTransfer(address _sender, uint256 _value, bytes _data);
 }
